Issue/2 criar tratamento excecao

+++ Service/Services/CargoService.cs
@@ -1,12 +1,7 @@
-﻿using Domain;
-using Ecclesia.Domain;
+﻿using Ecclesia.Domain;
 using Ecclesia.Repository.Contracts;
 using Ecclesia.Service.Contracts;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using System.Net;
 
 namespace Ecclesia.Service.Services
 {
@@ -21,6 +16,8 @@ namespace Ecclesia.Service.Services
 
         public async Task DeleteCargo(int id)
         {
+            var reg = _repository.GetCargo(id);
+            if (reg == null) throw new BusinessHttpResponseException(Messages.Message(HttpStatusCode.NotFound));
             await _repository.DeleteCargo(id);
         }
 
@@ -34,6 +31,7 @@ namespace Ecclesia.Service.Services
         public async Task<Cargo> GetCargo(int id)
         {
             var cargo = await _repository.GetCargo(id);
+            if (cargo ==  null) throw new BusinessHttpResponseException(HttpStatusCode.NotFound);
             return cargo;
         }
 
@@ -44,6 +42,8 @@ namespace Ecclesia.Service.Services
 
         public async Task UpdateCargo(Cargo cargo)
         {
+            var reg = _repository.GetCargo(cargo.Id);
+            if (reg == null) throw new BusinessHttpResponseException(Messages.Message(HttpStatusCode.NotFound));
             await _repository.UpdateCargo(cargo);
         }
     }
